python_mysql_dbinsert.py
from mysql.connector import MySQLConnection, Error
from python_mysql_dbconfig import read_db_config
 
def insert_user(name, password, email):
    query = "INSERT INTO users(username, password, email) " \
            "VALUES(%s,%s,%s)"

    args = (name, password, email)
 
    try:
        db_config = read_db_config()
        conn = MySQLConnection(**db_config)
 
        cursor = conn.cursor()
        cursor.execute(query, args)
 
        if cursor.lastrowid:
            print('last insert id', cursor.lastrowid)
        else:
            print('last insert id not found')
 
        conn.commit()
    except Error as error:
        print(error)
 
    finally:
        cursor.close()
        conn.close()