requirements.txt
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mercurial==3.3.post20150218
mysql-connector-python==2.0.4
vboxapi==1.0
virtualenv==13.1.2
Werkzeug==0.10.4
WTForms==2.0.2
